+++ Emergence/Contexts/App Setup/AuthViewController.swift
@@ -19,9 +19,4 @@ class AuthViewController: UIViewController {
 
     }
 
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-    
 }
